+++ eiknxx/projektai/models.py
@@ -5,7 +5,7 @@ import datetime
 class Projektas(models.Model):
     Project_token = models.ForeignKey(UserProfile,to_field='token',on_delete=models.PROTECT,null=True)
     Project_name = models.CharField(max_length=100)
-    Project_ID = models.IntegerField(primary_key=True)
+    Project_ID = models.BigIntegerField(primary_key=True)
     Parent_id = models.ForeignKey('self',to_field='Project_ID',null=True,blank=True,on_delete=models.CASCADE)
     Color = models.IntegerField(null=True,blank=True)
     Indent = models.IntegerField(null=True,blank=True)
@@ -16,7 +16,7 @@ class Projektas(models.Model):
 
 class Task(models.Model):
     task_Content = models.TextField(null=True,blank=True)
-    task_id = models.IntegerField(primary_key=True)
+    task_id = models.BigIntegerField(primary_key=True)
     task_project_id = models.ForeignKey(Projektas,to_field='Project_ID',on_delete=models.CASCADE)
     task_parent_id = models.ForeignKey('self',to_field='task_id',null=True,blank=True,on_delete=models.CASCADE)
     task_priority = models.IntegerField(null=True,blank=True)
